requirements.txt
numpy==1.23.4
Pillow==10.2.0
PyYAML==6.0.1
segmentation-models-pytorch==0.3.3
ultralytics>=8.0.225
